app/requirements.txt
amqp==5.3.1      
annotated-types==0.7.0      
anyio==4.9.0      
billiard==4.2.1      
celery==5.5.2      
click==8.2.0      
click-didyoumean==0.3.1      
click-plugins==1.1.1      
click-repl==0.3.0      
colorama==0.4.6      
dotenv==0.9.9      
fastapi==0.115.12   
filelock==3.18.0
flower==2.0.1
fsspec==2025.3.2   
h11==0.16.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.3
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
pillow==11.2.1
pip==25.1.1
prompt_toolkit==3.0.51
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==6.0.0
setuptools==65.5.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
uvicorn==0.34.2
vine==5.1.0
wcwidth==0.2.13